Update net structs + add client
Update the net structs to follow a KVRequest trait which provides a
Response associated type which is paired with the corresponding request type.
The client only deals with structs which implement the trait, while the
implementing types are packed into an enum which is passed over the wire
to the server.

Added an initial client which uses the trait definition and associated
types to implement the send_request logic in a generic way across each RPC